Add course data module and navbar dropdown
