+++ jest.config.js
@@ -1,8 +1,4 @@
-
 const config = {
-
-module.exports = {
-
   testEnvironment: 'jsdom',
   collectCoverage: true,
   coverageDirectory: 'coverage',
@@ -11,7 +7,6 @@ module.exports = {
   moduleNameMapper: {
     '\\.(css|less|scss|sass)$': 'identity-obj-proxy',
     '\\.(png|svg|jpg|jpeg|gif)$': '<rootDir>/__mocks__/fileMock.js'
-
   },
   transform: {
     '^.+\\.(js|jsx)$': ['@swc/jest', {
@@ -31,7 +26,3 @@ module.exports = {
 };
 
 export default config;
-
-  }
-};
-
