+++ setupTests.js
@@ -1,18 +1,5 @@
-
-require('@testing-library/jest-dom');
-
-
 import '@testing-library/jest-dom';
+import { TextEncoder, TextDecoder } from 'util';
 
-
-// Jest setup file for configuring the testing environment.
-// Currently imports utilities from React Testing Library.
-import '@testing-library/jest-dom/extend-expect';
-
-
-// Jest setup file for configuring the testing environment.
-// Currently imports utilities from React Testing Library.
-import '@testing-library/jest-dom/extend-expect';
-
-// Jest setup file
-
+global.TextEncoder = TextEncoder;
+global.TextDecoder = TextDecoder;
